+++ test/.eslintrc.js
@@ -1,0 +1,14 @@
+module.exports = {
+    env: {
+        mocha: true
+    },
+
+    globals: {
+        assert: false,
+        compareFixtures: false,
+        read: false,
+        processedCss: false,
+        postcss: false,
+        postcssUrl: false
+    }
+};